FIX AbstractMarkdownPlaceholderResolver was created and the required functions were moved from AbstractPlaceholderResolver NEW placeholders are implemented twice for references
